refactor: 패키지 구조 변경 및 삭제
