--- team3project/src/main/java/pack/give/util/FullBeanNameGenerator.java
@@ -1,13 +1,0 @@
-package pack.give.util;
-
-import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.beans.factory.support.BeanDefinitionRegistry;
-
-public class FullBeanNameGenerator implements org.springframework.beans.factory.support.BeanNameGenerator {
-
-    @Override
-    public String generateBeanName(BeanDefinition definition, BeanDefinitionRegistry registry) {
-        return definition.getBeanClassName();
-    }
-
-}
